requirements.txt
flask==3.1.0
flask-cors==5.0.1
supabase==2.11.0
python-dotenv==1.0.1
werkzeug==3.1.3
openai==1.58.1
scikit-learn==1.6.0
joblib==1.4.2
gunicorn==23.0.0
PyJWT==2.9.0
